+++ x/stakers/types/errors.go
@@ -17,7 +17,7 @@ var (
 	ErrValaddressSameAsStaker  = errors.Register(ModuleName, 1111, "Valaddress has same address as Valaddress")
 	ErrCanNotJoinDisabledPool  = errors.Register(ModuleName, 1112, "can not join disabled pool")
 	ErrInvalidIdentityString   = errors.Register(ModuleName, 1113, "invalid identity: %s")
-	ErrNotEnoughRewards        = errors.Register(ModuleName, 1114, "claim-rewards amount is larger than current rewards")
+	ErrNotEnoughRewards        = errors.Register(ModuleName, 1114, "claim amount is larger than current rewards")
 
 	ErrPoolLeaveAlreadyInProgress = errors.Register(ModuleName, 1117, "Pool leave is already in progress")
 	ErrValaccountUnauthorized     = errors.Register(ModuleName, 1118, "valaccount unauthorized")