+++ x/stakers/types/message_claim_commission_rewards.go
@@ -4,17 +4,36 @@ import (
 	"cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errorsTypes "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
-var _ sdk.Msg = &MsgClaimCommissionRewards{}
+var (
+	_ legacytx.LegacyMsg = &MsgClaimCommissionRewards{}
+	_ sdk.Msg            = &MsgClaimCommissionRewards{}
+)
 
-// GetSigners returns the expected signers for a MsgUpdateCommission message.
+func (msg *MsgClaimCommissionRewards) GetSignBytes() []byte {
+	bz := ModuleCdc.MustMarshalJSON(msg)
+	return sdk.MustSortJSON(bz)
+}
+
 func (msg *MsgClaimCommissionRewards) GetSigners() []sdk.AccAddress {
-	validator, _ := sdk.AccAddressFromBech32(msg.Creator)
+	validator, err := sdk.AccAddressFromBech32(msg.Creator)
+	if err != nil {
+		panic(err)
+	}
+
 	return []sdk.AccAddress{validator}
 }
 
-// ValidateBasic does a sanity check on the provided data.
+func (msg *MsgClaimCommissionRewards) Route() string {
+	return RouterKey
+}
+
+func (msg *MsgClaimCommissionRewards) Type() string {
+	return "kyve/stakers/MsgClaimCommissionRewards"
+}
+
 func (msg *MsgClaimCommissionRewards) ValidateBasic() error {
 	if _, err := sdk.AccAddressFromBech32(msg.Creator); err != nil {
 		return errors.Wrapf(errorsTypes.ErrInvalidAddress, "invalid validator address: %s", err)